+++ src/store/projects/types.js
@@ -1,0 +1,10 @@
+// Project types
+
+// Loading projects type
+export const PROJECTS_LOADING = 'PROJECTS_LOADING';
+
+// Successfully loaded projects type
+export const LOAD_PROJECTS_SUCCESS = 'LOAD_PROJECTS_SUCCESS';
+
+// Successfully loaded projects post type
+export const LOAD_PROJECT_POST_SUCCESS = 'LOAD_PROJECT_POST_SUCCESS';