+++ src/store/rootReducer.js
@@ -2,14 +2,14 @@ import {
     combineReducers
 } from 'redux'
 import base from './base/reducer'
+import projects from './projects/reducer'
 import blog from './blog/reducer'
-import work from './work/reducer'
 import team from './team/reducer'
 
 // Redux root reducers
 export const rootReducer = combineReducers({
     base,
+    projects,
     blog,
-    work,
     team
 })