+++ changelog.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.0.3
 
+### 2018-11-14
+
+#### Removed
+- Changed all intances of `work` to `project` for projects feature
+
 ### 2018-11-13
 
 #### Added